feat: update scopes to match Atlassian OAuth2 updates
There are two changes here.

First, we need to use and document the new fine-grained scopes used by
Atlassian cloud:

https://community.developer.atlassian.com/t/action-required-update-scopes-for-forge-and-oauth-2-0-3lo-apps/53299

Second, we switch to a different API to check for existence of issues.
Now instead of using Issue.Get we use Issue.GetTransitions. This
requires slightly fewer scope grants and still allows us to determine if
the issue exists or not.